# src/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"VendorProfileResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"businessName": {
						"type": "string"
					},
					"registrationNumber": {
						"type": "string"
					},
					"contactEmail": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"rejectionReason": {
						"type": "string",
						"nullable": true
					},
					"approvedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"businessName",
					"registrationNumber",
					"contactEmail",
					"country",
					"status",
					"rejectionReason",
					"approvedAt",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VendorOnboardingResponseDto": {
				"properties": {
					"vendor": {
						"$ref": "#/components/schemas/VendorProfileResponseDto"
					},
					"fee": {
						"properties": {
							"reference": {
								"type": "string"
							},
							"paymentUrl": {
								"type": "string"
							},
							"status": {
								"type": "string"
							},
							"currency": {
								"type": "string"
							},
							"amount": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"reference",
							"paymentUrl",
							"status",
							"currency",
							"amount"
						],
						"type": "object"
					}
				},
				"required": [
					"vendor",
					"fee"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VendorApplyDto": {
				"properties": {
					"businessName": {
						"type": "string",
						"minLength": 2,
						"maxLength": 200
					},
					"registrationNumber": {
						"type": "string",
						"minLength": 2,
						"maxLength": 100
					},
					"contactEmail": {
						"type": "string",
						"format": "email"
					},
					"country": {
						"type": "string",
						"description": "ISO 3166-1 alpha-2 country code",
						"minLength": 2,
						"maxLength": 2,
						"pattern": "^[A-Z]{2}$"
					},
					"currency": {
						"type": "string",
						"description": "Currency to pay the onboarding fee in",
						"minLength": 3,
						"maxLength": 3,
						"pattern": "^[A-Z]{3}$"
					},
					"callbackUrl": {
						"type": "string",
						"description": "URL Paystack redirects to after fee payment"
					}
				},
				"required": [
					"businessName",
					"registrationNumber",
					"contactEmail",
					"country",
					"currency"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VendorDocumentResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"documentType": {
						"type": "string"
					},
					"fileUrl": {
						"type": "string"
					},
					"verificationStatus": {
						"type": "string"
					},
					"uploadedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"documentType",
					"fileUrl",
					"verificationStatus",
					"uploadedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SubmitDocumentDto": {
				"properties": {
					"fileKey": {
						"type": "string"
					},
					"documentType": {
						"type": "string",
						"enum": [
							"business_registration",
							"tax_certificate",
							"id_proof",
							"other"
						]
					}
				},
				"required": [
					"fileKey",
					"documentType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AdminVendorListResponseDto": {
				"properties": {
					"data": {
						"items": {
							"$ref": "#/components/schemas/VendorProfileResponseDto"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					},
					"page": {
						"type": "number",
						"format": "double"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"total",
					"page",
					"limit",
					"totalPages"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ReviewVendorDto": {
				"properties": {
					"reason": {
						"type": "string",
						"minLength": 10,
						"maxLength": 500
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"PresignUploadResponseDto": {
				"properties": {
					"uploadUrl": {
						"type": "string"
					},
					"fileKey": {
						"type": "string"
					},
					"expiresIn": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"uploadUrl",
					"fileKey",
					"expiresIn"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PresignUploadDto": {
				"properties": {
					"filename": {
						"type": "string",
						"description": "Original filename including extension",
						"minLength": 1,
						"maxLength": 200
					},
					"contentType": {
						"type": "string",
						"description": "MIME type of the file",
						"minLength": 1,
						"maxLength": 100
					},
					"folder": {
						"type": "string",
						"enum": [
							"products",
							"documents",
							"avatars"
						],
						"description": "Where the file will be used"
					}
				},
				"required": [
					"filename",
					"contentType",
					"folder"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ecommerce-api",
		"version": "1.0.0",
		"description": "E-Commerce REST API — Node.js, Express, TypeScript, TypeORM, PostgreSQL",
		"contact": {}
	},
	"paths": {
		"/vendors/apply": {
			"post": {
				"operationId": "Apply",
				"responses": {
					"201": {
						"description": "Application submitted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorOnboardingResponseDto"
								}
							}
						}
					},
					"409": {
						"description": "Already has a vendor profile"
					}
				},
				"description": "Apply to become a vendor.\nReturns a paymentUrl — redirect the user there to pay the onboarding fee.\nStatus moves from pending_payment → pending_review after fee is confirmed.",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VendorApplyDto"
							}
						}
					}
				}
			}
		},
		"/vendors/me": {
			"get": {
				"operationId": "GetMyProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorProfileResponseDto"
								}
							}
						}
					},
					"404": {
						"description": "Vendor profile not found"
					}
				},
				"description": "Get the authenticated vendor's profile and status",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/vendors/documents": {
			"post": {
				"operationId": "SubmitDocument",
				"responses": {
					"201": {
						"description": "Document submitted",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorDocumentResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Fee payment required first"
					}
				},
				"description": "Submit a business document after paying the onboarding fee.\nUse POST /uploads/presign first to upload the file to S3,\nthen pass the returned fileKey here.",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubmitDocumentDto"
							}
						}
					}
				}
			}
		},
		"/vendors/me/documents": {
			"get": {
				"operationId": "GetMyDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VendorDocumentResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all documents submitted by the authenticated vendor",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/vendors/admin/pending": {
			"get": {
				"operationId": "GetPendingVendors",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AdminVendorListResponseDto"
								}
							}
						}
					},
					"403": {
						"description": "Admins only"
					}
				},
				"description": "Admin: list all vendors with pending_review status",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/vendors/{vendorId}/documents": {
			"get": {
				"operationId": "GetVendorDocuments",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/VendorDocumentResponseDto"
									},
									"type": "array"
								}
							}
						}
					},
					"403": {
						"description": "Admins only"
					}
				},
				"description": "Admin: view all documents submitted by a specific vendor",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vendorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vendors/{vendorId}/approve": {
			"post": {
				"operationId": "ApproveVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorProfileResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid status transition"
					},
					"404": {
						"description": "Vendor not found"
					}
				},
				"description": "Admin: approve a vendor application",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vendorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/vendors/{vendorId}/reject": {
			"post": {
				"operationId": "RejectVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorProfileResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Reason required"
					},
					"404": {
						"description": "Vendor not found"
					}
				},
				"description": "Admin: reject a vendor application with a written reason",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vendorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewVendorDto"
							}
						}
					}
				}
			}
		},
		"/vendors/{vendorId}/suspend": {
			"post": {
				"operationId": "SuspendVendor",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/VendorProfileResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid status transition"
					},
					"404": {
						"description": "Vendor not found"
					}
				},
				"description": "Admin: suspend an approved vendor",
				"tags": [
					"Vendors"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "vendorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ReviewVendorDto"
							}
						}
					}
				}
			}
		},
		"/uploads/presign": {
			"post": {
				"operationId": "Presign",
				"responses": {
					"200": {
						"description": "Pre-signed URL generated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PresignUploadResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "File type not allowed"
					}
				},
				"description": "Step 1 of 2 — Request a pre-signed S3 upload URL.\nThe client then PUTs the file directly to S3 using this URL.\nThe file never passes through the API server.",
				"tags": [
					"Uploads"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PresignUploadDto"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}